Split pathname to core and stdlib
