+++ test/stdlib/Pathname_test.rb
@@ -1,5 +1,5 @@
 require_relative "test_helper"
-require 'pathname'
+require "pathname" unless defined?(Pathname)
 
 class PathnameSingletonTest < Test::Unit::TestCase
   include TestHelper
@@ -314,17 +314,6 @@ class PathnameInstanceTest < Test::Unit::TestCase
                      Pathname('/unknown'), :file?
   end
 
-  def test_find
-    assert_send_type '() { (Pathname) -> untyped } -> nil',
-                     Pathname(__dir__), :find do end
-    assert_send_type '(ignore_error: bool) -> Enumerator[Pathname, nil]',
-                     Pathname(__dir__), :find, ignore_error: true
-    assert_send_type '(ignore_error: Symbol) -> Enumerator[Pathname, nil]',
-                     Pathname(__dir__), :find, ignore_error: :true
-    assert_send_type '() -> Enumerator[Pathname, nil]',
-                     Pathname(__dir__), :find
-  end
-
   def test_fnmatch
     assert_send_type '(String) -> bool',
                      Pathname('foo'), :fnmatch, 'fo*'
@@ -635,15 +624,6 @@ class PathnameInstanceTest < Test::Unit::TestCase
       target.mkdir
       assert_send_type '() -> 0',
                        target, :rmdir
-    end
-  end
-
-  def test_rmtree
-    Dir.mktmpdir do |dir|
-      target = Pathname(dir).join('target')
-      target.mkdir
-      assert_send_type '() -> Pathname',
-                       target, :rmtree
     end
   end
